Revise baseViewModel move collaboration source.
